// types/invoice.ts
export interface Contact {
  name: string
  email: string
  line1: string
  line2?: string
  town: string
  postcode: string
  country: string
}

export interface Client {
  name: string
  contact: Contact
}

export interface LineItem {
  title: string
  quantity: number
  price: number
  amount: number
}

export interface Invoice {
  id: number
  status: string
  date: string
  number: string
  client: Client
  supplier: Contact
  amount: 1000
  due: string
  line_items: LineItem[]
  total: number
  notes: string
  currency: string
  locale: string
}